dgram: setMulticastInterface() outgoing interface selection (review 3)
