+++ lib/dgram.js
@@ -566,8 +566,8 @@ Socket.prototype.setMulticastLoopback = function(arg) {
 Socket.prototype.setMulticastInterface = function(interfaceAddress) {
   this._healthCheck();
 
-  if (!interfaceAddress) {
-    throw new Error('interface address must be specified');
+  if (typeof interfaceAddress !== 'string') {
+    throw new TypeError('interfaceAddress must be a String');
   }
 
   const err = this._handle.setMulticastInterface(interfaceAddress);
